replaced code with common module
