+++ packages/adp-tooling/test/unit/base/prompt.test.ts
@@ -1,35 +1,7 @@
 import prompts from 'prompts';
-import { promptServiceKeys, promptCredentials, promptGeneratorInput } from '../../../src/base/prompt';
-import { join } from 'path';
+import { flexLayer, promptGeneratorInput } from '../../../src/base/prompt';
 
 describe('base/prompts', () => {
-    describe('promptCredentials', () => {
-        const username = '~user';
-        const password = '~pass';
-
-        test('no default provided', async () => {
-            prompts.inject([username, password]);
-            const creds = await promptCredentials();
-            expect(creds.username).toBe(username);
-            expect(creds.password).toBe(password);
-        });
-
-        test('default provided', async () => {
-            prompts.inject([undefined, password]);
-            const providedUser = '~other';
-            const creds = await promptCredentials(providedUser);
-            expect(creds.username).toBe(providedUser);
-            expect(creds.password).toBe(password);
-        });
-    });
-
-    describe('promptServiceKeys', () => {
-        test('Valid path provided.', async () => {
-            prompts.inject([join(__dirname, '../../../tsconfig.json')]);
-            expect(await promptServiceKeys()).toBeDefined();
-        });
-    });
-
     describe('promptGeneratorInput', () => {
         const defaults = {
             id: 'my.id',
@@ -38,11 +10,12 @@ describe('base/prompts', () => {
         };
 
         test('defaults provided', async () => {
+            prompts.inject([]);
             const config = await promptGeneratorInput(defaults);
             expect(config).toEqual({
                 app: {
-                    id: defaults.id,
-                    layer: undefined,
+                    id: `customer.${defaults.id}`,
+                    layer: flexLayer.CUSTOMER_BASE,
                     reference: defaults.reference,
                     title: undefined
                 },
@@ -53,12 +26,12 @@ describe('base/prompts', () => {
         });
 
         test('prompt everything', async () => {
-            prompts.inject(['VENDOR', defaults.id, defaults.reference, 'My Title', defaults.url]);
+            prompts.inject([flexLayer.VENDOR, defaults.id, defaults.reference, 'My Title', defaults.url]);
             const config = await promptGeneratorInput();
             expect(config).toEqual({
                 app: {
                     id: defaults.id,
-                    layer: 'VENDOR',
+                    layer: flexLayer.VENDOR,
                     reference: defaults.reference,
                     title: 'My Title'
                 },
