+++ packages/adp-tooling/tsconfig.json
@@ -1,10 +1,6 @@
 {
     "extends": "../../tsconfig.json",
-    "include": [
-        "../../types/ui5.d.ts",
-        "../../types/mem-fs-editor.d.ts",
-        "src"
-    ],
+    "include": ["../../types/ui5.d.ts", "../../types/mem-fs-editor.d.ts", "src"],
     "compilerOptions": {
         "rootDir": "src",
         "outDir": "dist"
@@ -24,6 +20,9 @@
         },
         {
             "path": "../store"
+        },
+        {
+            "path": "../system-access"
         },
         {
             "path": "../ui5-config"
